+++ src/OpenTelemetry.Api.ProviderBuilderExtensions/CHANGELOG.md
@@ -14,6 +14,12 @@
   created).
   ([#4508](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4508))
 
+* Added an `IServiceCollection.ConfigureOpenTelemetryMeterProvider` overload
+  which may be used to configure `MeterProviderBuilder`s while the
+  `IServiceCollection` is modifiable (before the `IServiceProvider` has been
+  created).
+  ([#XXXX](https://github.com/open-telemetry/opentelemetry-dotnet/pull/XXXX))
+
 ## 1.5.0-alpha.2
 
 Released 2023-Mar-31